Split useModelsTree and useCategoriesTree hooks
